requirements.txt
flask>=3.0
flask-cors>=4.0
flask-socketio>=5.3
requests>=2.31
python-dotenv>=1.0
neo4j>=5.0
